// src/app/models/abono.interface.ts
export interface AbonoI {
    id: number;
    urlqr: string;
    htmlSource: string;
    codigoAbono: string;
    localidad: string;
    asiento: string;
    nombres: string;
    cedula: string;
    correo: string;
    telefono: string;
    direccion: string;
    estado: string;
    tipo: string;
    observacion: string;
    registrosAbonos: any;
}